fix: align large-file threshold, dedupe redundant tempo map, seed memory monitor, test dpcm_converter (#334, #335, #336, #337)
PERF-14 (#334): move the advisory LARGE_FILE_THRESHOLD out of a bare inline
literal and into config, defaulting to MAX_PATTERN_EVENTS so the advisory
boundary finally matches the real parallel-detector sampling cap instead of
an unrelated 10000.

PERF-15 (#335): parse_midi_to_frames_with_analysis re-opened the MIDI file
and rebuilt an identical tempo map after parse_midi_to_frames had already
built one with the same config. Factor the shared open+tempo-map+event-parse
logic into _parse_frames_and_tempo_map so both entry points do it once.

PERF-16 (#336): MemoryMonitor.stop_monitoring() reported a misleading
peak_mb=0 when called before the background thread's first sample; seed
_memory_samples with an immediate RSS read in start_monitoring(). Swallowed
sampling-loop exceptions are now counted instead of silently discarded.

REG-18 (#337): add tests/test_dpcm_converter.py covering PCM conversion,
exact delta_encode/dpcm_compress byte output, and the 4081-byte DMC cap.
Writing the coverage surfaced a real gap: convert_wav_to_unsigned_pcm
returned float64 instead of uint8 for multi-channel 8-bit WAV input (the
mono-mixdown .mean() promotes to float and only the 16-bit path re-cast
afterward) -- fixed by casting back to the source dtype after mixdown.

+++ dpcm_sampler/dpcm_converter.py
@@ -12,9 +12,14 @@ def convert_wav_to_unsigned_pcm(wav_path, sample_rate=8000):
         dtype = {1: np.uint8, 2: np.int16}[sampwidth]
         data = np.frombuffer(frames, dtype=dtype)
 
-        # Convert to mono if needed
+        # Convert to mono if needed. Cast back to the source dtype
+        # immediately: .mean() always promotes to float64, and the 8-bit
+        # (sampwidth==1) path has no later re-cast to catch it -- only the
+        # sampwidth==2 branch below normalizes back to uint8, so a
+        # multi-channel 8-bit WAV previously left this function returning a
+        # float64 array where every other path returns uint8 (#337/REG-18).
         if channels > 1:
-            data = data.reshape(-1, channels).mean(axis=1)
+            data = data.reshape(-1, channels).mean(axis=1).astype(dtype)
 
         # Normalize to 8-bit unsigned PCM range
         if sampwidth == 2:
